% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_roc_curve.R
\name{adjusted_roc_curve}
\alias{adjusted_roc_curve}
\title{Obtain an adjusted ROC curve from rocdata}
\usage{
adjusted_roc_curve(
  rocdata = NULL,
  coef = NULL,
  test_grouped_dat = NULL,
  Xprotein_test = NULL,
  order = 1,
  refstate = NULL,
  verbose = T,
  py1,
  plot = TRUE
)
}
\arguments{
\item{rocdata}{NULL or an output from the rocdata function}

\item{coef}{a named vector for the fitted coefficient vector}

\item{test_grouped_dat}{a data table (sequence, labeled, unlabeled, seqId) containing validation examples}

\item{order}{a considered order of effects}

\item{refstate}{a character which will be used for the common reference state; the default is to use the most frequent amino acid as the reference state for each of the position.}

\item{verbose}{a logical value}

\item{py1}{the prevalence of positives in the unlabeled set (which will be used for a correction)}

\item{plot}{a logical value. If TRUE, a plot will be returned}
}
\value{
a list containing a data frame for generating a roc plot and the plot
}
\description{
Obtain an adjusted ROC curve from rocdata
}
