% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{auc}
\alias{auc}
\title{Return an adjusted (average) auc value from a fitted model at py1 ==py1[py1idx]}
\usage{
auc(vfit, py1idx = NULL)
}
\arguments{
\item{vfit}{vpudms.fit object}

\item{py1idx}{an index of py1 for auc to be returned at}
}
\description{
Return an adjusted (average) auc value from a fitted model at py1 ==py1[py1idx]
}
