% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrected_ROC_curve.R
\name{corrected_roc_curve}
\alias{corrected_roc_curve}
\title{Obtain a corrected ROC curve from rocdata}
\usage{
corrected_roc_curve(
  rocdata = NULL,
  coef = NULL,
  test_grouped_dat = NULL,
  order = 1,
  basestate = NULL,
  verbose = T,
  py1
)
}
\arguments{
\item{coef}{a named vector for the fitted coefficient vector}

\item{test_grouped_dat}{a data table (sequence, labeled, unlabeled, seqId) containing validation examples}

\item{order}{a considered order of effects}

\item{basestate}{a character value}

\item{verbose}{a logical value}

\item{py1}{the prevalence of positives in the unlabeled set (which will be used for a correction)}
}
\value{
a list containing a data frame for generating a roc plot and the plot
}
\description{
Obtain a corrected ROC curve from rocdata
}
