% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grouped_dat.R
\name{create_grouped_dat}
\alias{create_grouped_dat}
\title{Created a data set grouped by a sequence}
\usage{
create_grouped_dat(seq_l, seq_u)
}
\arguments{
\item{seq_l}{nl by 1 data.table (colname = V1)}

\item{seq_u}{nu by 1 data.table (colname = V1)}
}
\value{
A data table of (sequence, labeled, unlabeled, seqId)
}
\description{
Created a data set grouped by a sequence
}
