% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_frame.R
\name{create_model_frame}
\alias{create_model_frame}
\title{Create a model frame list from a grouped data set}
\usage{
create_model_frame(
  grouped_dat,
  order = 1,
  aggregate = T,
  basestate = NULL,
  verbose = T
)
}
\arguments{
\item{grouped_dat}{a grouped data set}

\item{order}{an integer of 1 or 2; main effects (order=1) or interaction effects (order=2) in X}

\item{aggregate}{a logical value}

\item{basestate}{a letter}

\item{verbose}{a logical value}
}
\value{
a list (X,z,wei,seqId,blockidx)
}
\description{
Create a model frame list from a grouped data set
}
