% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_protein_dat.R
\name{create_protein_dat}
\alias{create_protein_dat}
\title{Create a data set for protein}
\usage{
create_protein_dat(path_l, path_u, type = "sequences", WT = NULL)
}
\arguments{
\item{path_l}{the file path for the labeled sequences. File type can be txt or txt.gzip}

\item{path_u}{the file path for the unlabeled sequences}

\item{type}{type of the input files; "sequences" or "mutations" based on the type of the input; a default is "sequences"}

\item{WT}{a reference sequence if input type is "mutations"}
}
\value{
a data table containing (sequence, labeled, unlabeled, seqId)
}
\description{
Create a data set for protein
}
