% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective_n_seq.R
\name{effective_n_seq}
\alias{effective_n_seq}
\title{Number of effective sequences}
\usage{
effective_n_seq(n_reads, n_characterized)
}
\arguments{
\item{n_reads}{number of reads}

\item{n_characterized}{number of function characterizations}
}
\description{
Number of effective sequences
}
