% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_mut_less_than_thresh.R
\name{filter_mut_less_than_thresh}
\alias{filter_mut_less_than_thresh}
\title{Delete columns such that colSums <= thresh, and rows which contain a mutation into the deleted columns}
\usage{
filter_mut_less_than_thresh(
  Xprotein,
  thresh = 10,
  checkResFullRank = T,
  verbose = T
)
}
\arguments{
\item{Xprotein}{list (X,z,wei, seqId, blockId)}

\item{thresh}{an integer threshold}

\item{checkResFullRank}{a logical value}

\item{verbose}{a logical value. The default is TRUE}
}
\value{
a list (X,z,wei, seqId, blockId)
}
\description{
Delete columns such that colSums <= thresh, and rows which contain a mutation into the deleted columns
}
