% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pudms.R
\name{pudms}
\alias{pudms}
\title{DMS experiments analysis using PUlasso method}
\usage{
pudms(
  protein_dat,
  py1,
  order = 1,
  basestate = NULL,
  verbose = T,
  nobs_thresh = 10,
  lambda = 0,
  pvalue = T,
  intercept = F,
  maxit = 1000,
  p.adjust.method = "BH"
)
}
\arguments{
\item{protein_dat}{a data table containing (sequence, labeled, unlabeled, seqId)}

\item{order}{an integer}

\item{basestate}{a character for a single base state}

\item{verbose}{a logical value}

\item{nobs_thresh}{minimum required number of mutations}

\item{lambda}{l1 penalty}

\item{pvalue}{a logial value; if TRUE, p-values based on the asymptotic distribution are obtained}

\item{intercept}{a logical value; if TRUE, an estimated intercept is reported together with other coefficients}

\item{maxit}{maximum number of iterations}

\item{p.adjust.method}{method for multiple comparison}
}
\value{
a list containing a fit (from grpPUlasso), result_table (data.frame), and nobs
}
\description{
DMS experiments analysis using PUlasso method
}
