% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pudms.test.R
\name{pudms.test}
\alias{pudms.test}
\title{test a PU fit on a test data set}
\usage{
pudms.test(
  protein_dat,
  test_idx,
  nfolds = 10,
  pvalue = T,
  seed = round(runif(1, min = 1, max = 1000)),
  py1,
  maxit = 1000
)
}
\arguments{
\item{protein_dat}{a (full) grouped data set}

\item{test_idx}{an integer from 1 to nfolds, which will be used for a test set}

\item{nfolds}{the number of folds}

\item{pvalue}{a logical value, a default value is TRUE}

\item{seed}{a seed number for reproducibility}

\item{py1}{a prevalence P(y=1)}

\item{maxit}{max number of iterations}
}
\value{
a list containing
}
\description{
test a PU fit on a test data set
}
\examples{
library(puDMSdata)
data(grouped_DXS)
r = pudms.test(protein_dat = grouped_DXS, test_idx = 1, seed = 20122020,py1 = 0.1,pvalue = F )
}
