% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_pu.R
\name{pval_pu}
\alias{pval_pu}
\title{Obtain an asymptotic variance matrix of theta, standard errors, zvalues, and pvalues}
\usage{
pval_pu(X, z, theta, py1, weights = rep(1, nrow(X)), effective_n_prop = 1)
}
\arguments{
\item{X}{an n by p matrix}

\item{z}{an n by 1 vector}

\item{theta}{a p+1 by 1 vector}

\item{py1}{a numeric value}

\item{weights}{an n by 1 vector}

\item{effective_n_prop}{a numeric value}
}
\value{
a list (I = expected Fisher matrix, invI = inverse of I, se = sqrt(diag(invI)), zvalue, pvalue)
}
\description{
Obtain an asymptotic variance matrix of theta, standard errors, zvalues, and pvalues
}
