% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCdata.R
\name{rocdata}
\alias{rocdata}
\title{Create a data table (seqId, labeled, unlabeled, pr_test) to create a ROC curve}
\usage{
rocdata(coef, test_grouped_dat, order = 1, refstate = NULL, verbose = T)
}
\arguments{
\item{coef}{a named vector for the fitted coefficient vector}

\item{test_grouped_dat}{a data table (sequence, labeled, unlabeled, seqId) containing validation examples}

\item{order}{a considered order of effects}

\item{refstate}{a character which will be used for the common reference state; the default is to use the most frequent amino acid as the reference state for each of the position.}

\item{verbose}{a logical value}
}
\value{
a data table containing (seqId, labeled, unlabeled, pr_test)
}
\description{
Create a data table (seqId, labeled, unlabeled, pr_test) to create a ROC curve
}
