% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocker.R
\docType{data}
\name{rocker}
\alias{rocker}
\title{A rocker protein dataset}
\format{
A data frame with 79134 rows and 4 variables:
\describe{
  \item{\code{sequence}}{character; amino-acid sequences (of length 27)}
  \item{\code{labeled}}{integer; number of labeled sequences}
  \item{\code{unlabeled}}{integer; number of unlabeled sequences}
  \item{\code{seqId}}{integer; seqId from 1 to 79134} 
}
}
\usage{
rocker
}
\description{
A dataset for the protein rocker
}
\keyword{datasets}
