% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocplot.R
\name{rocplot}
\alias{rocplot}
\title{Draw an adjusted ROC curve}
\usage{
rocplot(roc_curve, py1)
}
\arguments{
\item{roc_curve}{an roc curve from the output of the function adjusted_roc_curve()}

\item{py1}{a numeric value}
}
\description{
Draw an adjusted ROC curve
}
