% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.model.matrix.parallel.R
\name{sparse.model.matrix.parallel}
\alias{sparse.model.matrix.parallel}
\title{Create a sparse model matrix in a parallel manner}
\usage{
sparse.model.matrix.parallel(order = c(1, 2), data, nCores = 1, verbose = TRUE)
}
\arguments{
\item{order}{an integer; 1= main effects, 2= main effects + pairwise effects}

\item{data}{data to build a model matrix from}

\item{nCores}{number of threads for parallel computing}
}
\description{
Create a sparse model matrix in a parallel manner
}
