% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.pudms.R
\name{v.pudms}
\alias{v.pudms}
\title{test a PU fit on a test data set}
\usage{
v.pudms(
  protein_dat,
  py1 = NULL,
  nhyperparam = 10,
  nfolds = 5,
  test_idx = 1:nfolds,
  seed = round(runif(1, min = 1, max = 1000)),
  order = 1,
  refstate = NULL,
  verbose = T,
  nobs_thresh = 10,
  lambda = 0,
  pvalue = FALSE,
  n_eff_prop = 1,
  intercept = F,
  maxit = 1000,
  eps = 0.001,
  inner_eps = 0.01,
  initial_coef = NULL,
  p.adjust.method = "BH",
  is.smooth.roc = TRUE,
  tol = 1e-05,
  nCores = 1,
  full.fit = FALSE,
  full.fit.pvalue = FALSE,
  outfile = NULL
)
}
\arguments{
\item{protein_dat}{input data. A data table containing (sequence, labeled, unlabeled, seqId)}

\item{py1}{a numeric value, a numeric vector or NULL; the prevalence of positives in the unlabeled data. If length(py1) >1, optimal py1 will be chosen based on auc values on a test data set. If NULL (default), a sequence of py1 values (of length nhyperparam)--ranging from 0.001 to 0.5 interpolated in a log scale--will be considered.}

\item{nhyperparam}{an integer for the length of the py1 sequence if py1 == NULL}

\item{nfolds}{the number of subsamples. (nfolds -1)/nfolds splits will be used for training, and the rest will be used for testing.}

\item{test_idx}{a vector of indices of cross-validation models which will be fitted. Default is to fit the model for each of the cross-validation fold.}

\item{seed}{a seed number for reproducibility}

\item{order}{an integer; 1= main effects, 2= main effects + pairwise effects}

\item{refstate}{a character which will be used for the common reference state; the default is to use the most frequent amino acid as the reference state for each of the position.}

\item{verbose}{a logical value. The default is TRUE}

\item{nobs_thresh}{the number of minimum required mutations per position}

\item{lambda}{l1 penalty}

\item{pvalue}{a logial value; if TRUE, p-values based on the asymptotic distribution are obtained}

\item{n_eff_prop}{proportion of an effective sample size}

\item{intercept}{a logical value; if TRUE, an estimated intercept is reported together with other coefficients}

\item{maxit}{maximum number of iterations}

\item{eps}{convergence threshold for the outer loop}

\item{inner_eps}{convergence threshold for the inner loop}

\item{initial_coef}{a vector representing an initial point where we start PUlasso algorithm from.}

\item{p.adjust.method}{method for multiple comparison}

\item{is.smooth.roc}{a logical value; if TRUE (default), a smoothed roc curve will be used in checking whether the roc curve is contained by the maximal roc curve at each py1}

\item{tol}{a numeric value; if the roc estimated curve <= y+tol, the estimated roc curve is determined to be contained by the maximal curve.}

\item{nCores}{the number of threads for computing.}

\item{full.fit}{a logical value; if TRUE, the model will be fitted using a full data set and at a chosen py1.}

\item{full.fit.pvalue}{a logical value; if TRUE, p-values for the full fit will be returned}

\item{outfile}{NULL or a string; if a string is provided, an output with the name of the string will be exported in a working directory.}
}
\value{
a list containing v.dmsfit (all fits using training/test splits), roc_curves (average roc curve at each py1), dmsfit (pudms.fit using a full data set at the selected py1),  folds (test/training split information), py1 (a sequence of py1 values used for searching), py1.opt (the selected py1 value based on the predictive performance of the models)
}
\description{
test a PU fit on a test data set
}
