% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald_pvalue.R
\name{wald_pvalue}
\alias{wald_pvalue}
\title{Calculate a Wald statistic}
\usage{
wald_pvalue(
  coef,
  vcov,
  position,
  excludeStates = "*",
  order = 1,
  position2 = NULL,
  verbose = TRUE
)
}
\arguments{
\item{coef}{an estimated parameter}

\item{vcov}{a variance-covariance matrix}

\item{verbose}{a logical value.}

\item{idx}{indices for a group which we want to extract a p-value for}
}
\description{
Calculate a Wald statistic
}
