% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_frame.R
\name{create_model_frame}
\alias{create_model_frame}
\title{Create a model frame list from a grouped data set}
\usage{
create_model_frame(
  grouped_dat,
  order = 1,
  aggregate = T,
  refstate = NULL,
  verbose = T
)
}
\arguments{
\item{grouped_dat}{a grouped data set}

\item{order}{an integer; 1= main effects, 2= main effects + pairwise effects}

\item{aggregate}{a logical value}

\item{refstate}{a character which will be used for the common reference state; the default is to use the most frequent amino acid as the reference state for each of the position.}

\item{verbose}{a logical value}
}
\value{
a list (X,z,wei,seqId,blockidx)
}
\description{
Create a model frame list from a grouped data set
}
