% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROCdata.R
\name{ROCdata}
\alias{ROCdata}
\title{Create a data table (seqId, labeled, unlabeled, pr_test) to create a ROC curve}
\usage{
ROCdata(coef, test_grouped_dat, order = 1, basestate = NULL, verbose = T)
}
\arguments{
\item{coef}{a named vector for the fitted coefficient vector}

\item{test_grouped_dat}{a data table (sequence, labeled, unlabeled, seqId) containing validation examples}

\item{order}{a considered order of effects}

\item{verbose}{a logical value}
}
\value{
a data table containing (seqId, labeled, unlabeled, pr_test)
}
\description{
Create a data table (seqId, labeled, unlabeled, pr_test) to create a ROC curve
}
